refactor: separate UI from transaction authorization
Fix: #823
